Big cleanup of welcome and service worker managers
