--- typescript/web/src/components/app-lifecycle-manager/index.ts
@@ -1,1 +1,0 @@
-export { AppLifecycleManager } from "./app-lifecycle-manager";
