--- typescript/web/src/components/app-lifecycle-manager/welcome-modal/index.ts
@@ -1,1 +1,0 @@
-export { WelcomeModal } from "./welcome-modal";
