+++ typescript/web/src/components/service-worker-manager/index.ts
@@ -1,0 +1,2 @@
+export { ServiceWorkerManagerModal } from "./service-worker-manager-modal";
+export { ServiceWorkerManagerBackground } from "./service-worker-manager-background";
