+++ typescript/web/src/components/service-worker-manager/service-worker-manager-modal.tsx
@@ -3,7 +3,6 @@ import { useErrorHandler } from "react-error-boundary";
 import type { Workbox } from "workbox-window";
 import { useQueryParam, StringParam } from "use-query-params";
 import { UpdateServiceWorkerModal } from "./update-service-worker-modal/update-service-worker-modal";
-import { WelcomeModal } from "./welcome-modal";
 
 declare global {
   interface Window {
@@ -11,11 +10,7 @@ declare global {
   }
 }
 
-export const AppLifecycleManager = ({
-  noModals = false,
-}: {
-  noModals?: boolean; // Should be true for website and pages where you dont want service worker related modals to appear
-}) => {
+export const ServiceWorkerManagerModal = () => {
   // See https://docs.cypress.io/guides/core-concepts/conditional-testing#Welcome-wizard
   // This param can have several values:
   //   - undefined: Normal behavior, only show the update modal when needed
@@ -74,33 +69,6 @@ export const AppLifecycleManager = ({
       return;
     }
 
-    if (noModals) {
-      // Don't show any modals but register and update service worker automatically
-      // Good for website pages
-      try {
-        const wb = window.workbox;
-
-        const updateServiceWorkerWhenWaiting = () => {
-          updateServiceWorker();
-          wb.removeEventListener("waiting", updateServiceWorkerWhenWaiting);
-        };
-
-        wb.addEventListener("waiting", updateServiceWorkerWhenWaiting);
-
-        // never forget to call register as auto register is turned off in next.config.js
-        wb.register();
-
-        // eslint-disable-next-line consistent-return
-        return () => {
-          wb.removeEventListener("waiting", updateServiceWorkerWhenWaiting);
-        };
-      } catch (error) {
-        // eslint-disable-next-line no-console
-        console.warn("This browser does not support service worker");
-      }
-    }
-
-    // If !noModals then we show the full service worker modals
     try {
       const wb = window.workbox;
 
@@ -142,19 +110,13 @@ export const AppLifecycleManager = ({
     } catch (error) {
       handleError(error);
     }
-  }, [noModals]);
+  }, []);
 
-  if (noModals) {
-    return null;
-  }
   return (
-    <>
-      <WelcomeModal />
-      <UpdateServiceWorkerModal
-        isOpen={isUpdateServiceWorkerModalOpen}
-        onClose={closeUpdateServiceWorkerModal}
-        onConfirm={updateServiceWorker}
-      />
-    </>
+    <UpdateServiceWorkerModal
+      isOpen={isUpdateServiceWorkerModalOpen}
+      onClose={closeUpdateServiceWorkerModal}
+      onConfirm={updateServiceWorker}
+    />
   );
 };
