+++ typescript/web/src/components/service-worker-manager/update-service-worker-modal/index.ts
