+++ typescript/web/src/components/service-worker-manager/update-service-worker-modal/update-service-worker-modal.tsx
