+++ typescript/web/src/components/welcome-manager/index.ts
@@ -1,0 +1,1 @@
+export { WelcomeModal as WelcomeManager } from "./welcome-modal";
