+++ typescript/web/src/components/welcome-manager/welcome-modal.tsx
@@ -12,7 +12,7 @@ import { QueryParamConfig, StringParam, useQueryParam } from "use-query-params";
 import type { Dataset as DatasetType } from "@labelflow/graphql-types";
 import { useRouter, NextRouter } from "next/router";
 import { useErrorHandler } from "react-error-boundary";
-import { browser } from "../../../utils/detect-scope";
+import { browser } from "../../utils/detect-scope";
 import { BrowserWarning } from "./steps/browser-warning";
 import { BrowserError } from "./steps/browser-error";
 import { Welcome } from "./steps/welcome";
@@ -20,7 +20,7 @@ import { Loading } from "./steps/loading";
 import {
   checkServiceWorkerReady,
   messageNoWindow,
-} from "../../../utils/check-service-worker";
+} from "../../utils/check-service-worker";
 
 const tutorialDatasetFirstImageUrl =
   "/local/datasets/tutorial-dataset/images/2bbbf664-5810-4760-a10f-841de2f35510";
