+++ typescript/web/src/pages/auth/signin.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from "react";
 import { useRouter } from "next/router";
 import { useQueryParam, StringParam } from "use-query-params";
 import { Meta } from "../../components/meta";
-import { AppLifecycleManager } from "../../components/app-lifecycle-manager";
+import { ServiceWorkerManagerBackground } from "../../components/service-worker-manager";
 import { Layout } from "../../components/layout";
 import { SigninModal } from "../../components/auth-manager/signin-modal";
 
@@ -15,7 +15,7 @@ const LocalDatasetsIndexPage = () => {
   }, []);
   return (
     <>
-      <AppLifecycleManager noModals />
+      <ServiceWorkerManagerBackground />
       <Meta title="LabelFlow | Sign in" />
       <Layout>
         <SigninModal
