+++ typescript/web/src/pages/graphiql.tsx
@@ -12,7 +12,8 @@ import {
 import { RiArrowRightSLine } from "react-icons/ri";
 import { Meta } from "../components/meta";
 import { Layout } from "../components/layout";
-import { AppLifecycleManager } from "../components/app-lifecycle-manager";
+import { ServiceWorkerManagerModal } from "../components/service-worker-manager";
+import { WelcomeManager } from "../components/welcome-manager";
 import { AuthManager } from "../components/auth-manager";
 
 const GraphiQL = dynamic(() => import("../components/graphiql"), {
@@ -30,7 +31,8 @@ const GraphiQL = dynamic(() => import("../components/graphiql"), {
 const GraphqlPlayground = () => {
   return (
     <>
-      <AppLifecycleManager />
+      <ServiceWorkerManagerModal />
+      <WelcomeManager />
       <Meta title="LabelFlow | GraphiQL" />
       <AuthManager />
       <Layout
