+++ typescript/web/src/pages/local/datasets/[datasetSlug]/index.tsx
@@ -15,13 +15,14 @@ import { useErrorHandler } from "react-error-boundary";
 import { RiArrowRightSLine } from "react-icons/ri";
 import NextLink from "next/link";
 import { Meta } from "../../../../components/meta";
-import { AppLifecycleManager } from "../../../../components/app-lifecycle-manager";
+import { ServiceWorkerManagerModal } from "../../../../components/service-worker-manager";
 import { Layout } from "../../../../components/layout";
 import Error404Page from "../../../404";
 import { ExportButton } from "../../../../components/export-button";
 import { ImportButton } from "../../../../components/import-button";
 import { KeymapButton } from "../../../../components/layout/top-bar/keymap-button";
 import { AuthManager } from "../../../../components/auth-manager";
+import { WelcomeManager } from "../../../../components/welcome-manager";
 
 const getDataset = gql`
   query getDataset($slug: String!) {
@@ -64,7 +65,7 @@ const DatasetIndexPage = () => {
     }
     return (
       <>
-        <AppLifecycleManager />
+        <ServiceWorkerManagerModal />
         <Error404Page />
       </>
     );
@@ -72,7 +73,8 @@ const DatasetIndexPage = () => {
 
   return (
     <>
-      <AppLifecycleManager />
+      <ServiceWorkerManagerModal />
+      <WelcomeManager />
       <AuthManager />
       <Meta title={`LabelFlow | ${datasetName ?? "Dataset"}`} />
       <Layout
