+++ typescript/web/src/pages/posts/index.tsx
@@ -5,7 +5,7 @@ import { NavBar } from "../../components/website/Navbar/NavBar";
 import { Footer } from "../../components/website/Footer/Footer";
 import { Meta } from "../../components/meta";
 import { ArticlesList } from "../../components/website/Blog/articles-list";
-import { AppLifecycleManager } from "../../components/app-lifecycle-manager";
+import { ServiceWorkerManagerBackground } from "../../components/service-worker-manager";
 
 export default function Posts({
   previewArticles,
@@ -14,7 +14,7 @@ export default function Posts({
 }) {
   return (
     <>
-      <AppLifecycleManager noModals />
+      <ServiceWorkerManagerBackground />
       <Meta title="LabelFlow | Blog" />
       <Box minH="640px">
         <NavBar />
