+++ typescript/web/src/pages/pricing.tsx
@@ -7,12 +7,12 @@ import { NavBar } from "../components/website/Navbar/NavBar";
 import { Footer } from "../components/website/Footer/Footer";
 // import { PricingTable } from "../components/website/Pricing/PricingTable";
 import { Meta } from "../components/meta";
-import { AppLifecycleManager } from "../components/app-lifecycle-manager";
+import { ServiceWorkerManagerBackground } from "../components/service-worker-manager";
 
 export default function Pricing() {
   return (
     <>
-      <AppLifecycleManager noModals />
+      <ServiceWorkerManagerBackground />
       <Meta title="LabelFlow | Pricing" />
       <Box minH="640px">
         <NavBar />
