+++ typescript/web/src/pages/thank-you.tsx
@@ -6,12 +6,12 @@ import { NavBar } from "../components/website/Navbar/NavBar";
 import { Survey } from "../components/website/Survey/Survey";
 import { Footer } from "../components/website/Footer/Footer";
 import { Meta } from "../components/meta";
-import { AppLifecycleManager } from "../components/app-lifecycle-manager";
+import { ServiceWorkerManagerBackground } from "../components/service-worker-manager";
 
 export default function ThankYou() {
   return (
     <>
-      <AppLifecycleManager noModals />
+      <ServiceWorkerManagerBackground />
       <Meta title="LabelFlow | Thank you" />
       <Box minH="640px">
         <NavBar />
