+++ typescript/web/src/pages/workspaces/local.tsx
@@ -1,9 +1,10 @@
 import { Spinner, Center } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import React, { useEffect } from "react";
-import { AppLifecycleManager } from "../../components/app-lifecycle-manager";
+import { ServiceWorkerManagerModal } from "../../components/service-worker-manager";
 import { AuthManager } from "../../components/auth-manager";
 import { Layout } from "../../components/layout";
+import { WelcomeManager } from "../../components/welcome-manager";
 
 const LocalWorkspacesRedirectPage = () => {
   const router = useRouter();
@@ -14,7 +15,8 @@ const LocalWorkspacesRedirectPage = () => {
 
   return (
     <>
-      <AppLifecycleManager />
+      <ServiceWorkerManagerModal />
+      <WelcomeManager />
       <AuthManager />
       <Layout>
         <Center h="full">
